+++ poe/rspoe.h
@@ -13,7 +13,7 @@ RSPOE_EXPORT PoeState getPortState(int port);
 RSPOE_EXPORT int setPortState(int port, PoeState state);
 RSPOE_EXPORT float getPortVoltage(int port);
 RSPOE_EXPORT float getPortCurrent(int port);
-RSPOE_EXPORT int getPortPower(int port);
+RSPOE_EXPORT float getPortPower(int port);
 RSPOE_EXPORT const char* getLastPoeError();
 
 #ifdef __cplusplus
